refactor(update): separate convergence boundaries
