+++ apps/cli/src/commands/update/report.ts
@@ -1,15 +1,18 @@
-import {
-  type UpdateCommandReport,
-  UpdateCommandReportSchema,
-  type UpdateConvergencePlan,
-} from "@station/contracts";
+import { type UpdateCommandReport, UpdateCommandReportSchema } from "@station/contracts";
 import { shellQuote } from "@station/runtime";
 import type { CliRunResult } from "../../cliTypes.js";
-import { sanitizePublicUpdateReport } from "../../update/publicUpdateReport.js";
+import { sanitizePublicUpdateReport } from "../../update/publicUpdateReportAdapter.js";
+import { updateCommandExitCode } from "../../update/updateCommandStatusPolicy.js";
+import { encodeUpdateTerminalText } from "../../update/updateTerminalText.js";
 import type { UpdateRequest } from "./args.js";
 
 export type { UpdateCommandReport } from "@station/contracts";
 
+/**
+ * ADAPTER
+ *
+ * Presents one strict public update report as JSON or terminal-safe text with its process status.
+ */
 export function updateCommandResult(
   reportInput: UpdateCommandReport,
   output: UpdateRequest["output"],
@@ -23,39 +26,16 @@ export function updateCommandResult(
   };
 }
 
-/**
- * POLICY
- *
- * Maps the strict update disposition to the CLI process contract shared by the presenter and
- * successor-child boundary.
- */
-export function updateCommandExitCode(report: Pick<UpdateCommandReport, "status">): 0 | 1 {
-  switch (report.status) {
-    case "current":
-    case "updated":
-    case "planned":
-    case "deferred":
-      return 0;
-    case "failed":
-    case "blocked":
-    case "reap-required":
-    case "intentionally-incomplete":
-      return 1;
-  }
-}
-
-export function nonExecutedPhases(plan: UpdateConvergencePlan) {
-  return plan.phases.map((phase) => ({ id: phase.id, status: "not-executed" as const }));
-}
+export { updateCommandExitCode } from "../../update/updateCommandStatusPolicy.js";
 
 function renderUpdateReport(report: UpdateCommandReport): string {
   const plan = report.initial.plan;
   const lines = [
-    `channel: ${report.channel}`,
-    `status: ${report.status}`,
+    `channel: ${encodeUpdateTerminalText(report.channel)}`,
+    `status: ${encodeUpdateTerminalText(report.status)}`,
     `artifact installed: ${artifactText(report.current)}`,
     `artifact selected: ${artifactText(report.target)}`,
-    `artifact application: ${report.artifactApplication.status}`,
+    `artifact application: ${encodeUpdateTerminalText(report.artifactApplication.status)}`,
   ];
   if (
     (report.artifactApplication.status === "preview" ||
@@ -64,38 +44,55 @@ function renderUpdateReport(report: UpdateCommandReport): string {
   ) {
     lines.push(
       `manager command: ${report.artifactApplication.managerCommand
-        .map((value) => shellQuote(value))
+        .map((value) => shellQuote(encodeUpdateTerminalText(value)))
         .join(" ")}`,
     );
   }
   lines.push(
-    `plan evaluator: ${report.initial.evaluator}`,
-    `plan digest: ${plan.digest.value}`,
-    `plan status: ${plan.status}`,
+    `plan evaluator: ${encodeUpdateTerminalText(report.initial.evaluator)}`,
+    `plan digest: ${encodeUpdateTerminalText(plan.digest.value)}`,
+    `plan status: ${encodeUpdateTerminalText(plan.status)}`,
     "hooks:",
   );
   if (plan.components.hooks.length === 0) lines.push("  none configured");
   for (const hook of plan.components.hooks) {
-    lines.push(`  ${safeText(hook.provider)}: ${hook.action} (${hook.reason})`);
+    lines.push(
+      `  ${encodeUpdateTerminalText(hook.provider)}: ${encodeUpdateTerminalText(hook.action)} (${encodeUpdateTerminalText(hook.reason)})`,
+    );
   }
   lines.push(
-    `observer: ${plan.components.observer.action} (${plan.components.observer.reason})`,
-    `host: ${plan.components.host.action} (${plan.components.host.reason})`,
-    `terminals: ${plan.components.terminals.action} (${plan.components.terminals.reason}); live=${plan.components.terminals.liveCount} recoverable=${plan.components.terminals.recoverableCount} non-resumable=${plan.components.terminals.nonResumableCount} unknown=${plan.components.terminals.unknownRecoveryCount}`,
-    `recovery: ${plan.components.recovery.relevance}/${plan.components.recovery.status}`,
-    `reconcile: ${plan.components.reconcile.action} (${plan.components.reconcile.reason})`,
+    `observer: ${encodeUpdateTerminalText(plan.components.observer.action)} (${encodeUpdateTerminalText(plan.components.observer.reason)})`,
+    `host: ${encodeUpdateTerminalText(plan.components.host.action)} (${encodeUpdateTerminalText(plan.components.host.reason)})`,
+    `terminals: ${encodeUpdateTerminalText(plan.components.terminals.action)} (${encodeUpdateTerminalText(plan.components.terminals.reason)}); live=${plan.components.terminals.liveCount} recoverable=${plan.components.terminals.recoverableCount} non-resumable=${plan.components.terminals.nonResumableCount} unknown=${plan.components.terminals.unknownRecoveryCount}`,
+    `recovery: ${encodeUpdateTerminalText(plan.components.recovery.relevance)}/${encodeUpdateTerminalText(plan.components.recovery.status)}`,
+    `reconcile: ${encodeUpdateTerminalText(plan.components.reconcile.action)} (${encodeUpdateTerminalText(plan.components.reconcile.reason)})`,
     "ordered convergence phases:",
   );
   for (const phase of plan.phases) {
-    lines.push(`  ${phase.id}: ${phase.action} (${phase.reason})`);
+    lines.push(
+      `  ${encodeUpdateTerminalText(phase.id)}: ${encodeUpdateTerminalText(phase.action)} (${encodeUpdateTerminalText(phase.reason)})`,
+    );
   }
-  lines.push(`result: ${report.result.kind}`);
+  lines.push(`result: ${encodeUpdateTerminalText(report.result.kind)}`);
+  const resultPhases = nonMutatingResultPhases(report);
+  if (resultPhases !== undefined) {
+    lines.push("result convergence phases:");
+    for (const phase of resultPhases) {
+      lines.push(
+        `  ${encodeUpdateTerminalText(phase.id)}: ${encodeUpdateTerminalText(phase.status)}`,
+      );
+    }
+  }
   for (const audit of actionAudits(report)) {
-    lines.push(`executed digest: ${audit.planDigest} by ${audit.executor}`);
+    lines.push(
+      `executed digest: ${encodeUpdateTerminalText(audit.planDigest)} by ${encodeUpdateTerminalText(audit.executor)}`,
+    );
     for (const action of audit.actions) {
       lines.push(
-        `  ${action.phase}: ${action.action} ${action.status}${
-          action.provider === undefined ? "" : ` provider=${safeText(action.provider)}`
+        `  ${encodeUpdateTerminalText(action.phase)}: ${encodeUpdateTerminalText(action.action)} ${encodeUpdateTerminalText(action.status)}${
+          action.provider === undefined
+            ? ""
+            : ` provider=${encodeUpdateTerminalText(action.provider)}`
         }`,
       );
     }
@@ -103,25 +100,35 @@ function renderUpdateReport(report: UpdateCommandReport): string {
   const final = finalEvidence(report);
   if (final !== undefined) {
     lines.push(
-      `verified plan: ${final.plan.digest.value} (${final.plan.status}) by ${final.evaluator}`,
+      `verified plan: ${encodeUpdateTerminalText(final.plan.digest.value)} (${encodeUpdateTerminalText(final.plan.status)}) by ${encodeUpdateTerminalText(final.evaluator)}`,
     );
   }
-  for (const warning of report.warnings) lines.push(`warning: ${warning.message}`);
+  for (const warning of report.warnings) {
+    lines.push(`warning: ${encodeUpdateTerminalText(warning.message)}`);
+  }
   if (report.error !== undefined)
-    lines.push(`error: ${report.error.message} (${report.error.code})`);
+    lines.push(
+      `error: ${encodeUpdateTerminalText(report.error.message)} (${encodeUpdateTerminalText(report.error.code)})`,
+    );
   if (report.cause !== undefined)
-    lines.push(`cause: ${safeText(report.cause.message)} (${safeText(report.cause.code)})`);
+    lines.push(
+      `cause: ${encodeUpdateTerminalText(report.cause.message)} (${encodeUpdateTerminalText(report.cause.code)})`,
+    );
   if (report.startupEvidence !== undefined) {
     lines.push("observer startup evidence:");
-    lines.push(`  boot log: ${safeText(report.startupEvidence.bootLogPath)}`);
+    lines.push(`  boot log: ${encodeUpdateTerminalText(report.startupEvidence.bootLogPath)}`);
     if (report.startupEvidence.bootLogTail !== undefined) {
-      lines.push(`  bounded boot log tail: ${safeText(report.startupEvidence.bootLogTail)}`);
+      lines.push(
+        `  bounded boot log tail: ${encodeUpdateTerminalText(report.startupEvidence.bootLogTail)}`,
+      );
     }
   }
   if (report.recoveryCommands.length > 0) {
     lines.push("recovery commands:");
     for (const command of report.recoveryCommands) {
-      lines.push(`  ${command.map((value) => shellQuote(value)).join(" ")}`);
+      lines.push(
+        `  ${command.map((value) => shellQuote(encodeUpdateTerminalText(value))).join(" ")}`,
+      );
     }
   }
   return `${lines.join("\n")}\n`;
@@ -160,20 +167,22 @@ function finalEvidence(report: UpdateCommandReport) {
   }
 }
 
-function artifactText(artifact: UpdateCommandReport["current"]): string {
-  return artifact.revision === undefined
-    ? safeText(artifact.version)
-    : `${safeText(artifact.version)} (${safeText(artifact.revision)})`;
+function nonMutatingResultPhases(report: UpdateCommandReport) {
+  switch (report.result.kind) {
+    case "preview":
+    case "deferred":
+    case "non-mutating-stop":
+      return report.result.phases;
+    case "already-converged":
+    case "current-runtime-execution":
+    case "successor-runtime-execution":
+    case "execution-failed":
+      return undefined;
+  }
 }
 
-function safeText(value: string): string {
-  let escaped = "";
-  for (const character of value) {
-    const point = character.codePointAt(0) ?? 0;
-    escaped +=
-      point <= 31 || (point >= 127 && point <= 159)
-        ? `\\u${point.toString(16).padStart(4, "0")}`
-        : character;
-  }
-  return escaped;
+function artifactText(artifact: UpdateCommandReport["current"]): string {
+  return artifact.revision === undefined
+    ? encodeUpdateTerminalText(artifact.version)
+    : `${encodeUpdateTerminalText(artifact.version)} (${encodeUpdateTerminalText(artifact.revision)})`;
 }
