+++ apps/cli/src/update/updateConvergencePort.ts
@@ -1,0 +1,21 @@
+import type { HostHandoffFidelity, UpdateArtifact, UpdateCommandReport } from "@station/contracts";
+import type { UpdateChannelId } from "./updateChannel.js";
+
+export type UpdateConvergenceRequest = {
+  channel?: UpdateChannelId;
+  mode: "preview" | "apply";
+  packageManager: "defer" | "drive";
+  handoff?: HostHandoffFidelity;
+  reap: boolean;
+  evaluator: "incumbent-cli" | "successor-cli";
+  successorTarget?: UpdateArtifact;
+};
+
+/**
+ * DRIVING PORT
+ *
+ * Offers exact-target update convergence to CLI actors without exposing its driven capabilities.
+ */
+export interface UpdateConvergencePort {
+  run(request: UpdateConvergenceRequest): Promise<UpdateCommandReport>;
+}
