+++ docs/observer-architecture.md
@@ -236,6 +236,7 @@ areas contain the following responsibilities:
 | `diagnostics/` | doctor and diagnostic collection, the local-evidence port, and local representation translation | Diagnostic use cases aggregate core, journal, persistence-health, provider, configuration, and typed local evidence; `localEvidenceSource.ts` alone owns state, JSONL log, and hook-spool filesystem traversal. |
 | `features/` | feature-flag evaluation | Deterministic application policy. |
 | `apps/cli/src/observerProviders.ts` | concrete provider construction and role assignment | Outer composition root. |
+| `apps/cli/src/commands/update.ts`, `apps/cli/src/update/updateConvergenceUseCase.ts`, update ports/adapters, and `apps/cli/src/commands/registry/update.ts` | update argument translation, convergence orchestration, strict child transport, public report projection, and default wiring | The command is a driving adapter over the update convergence port. The use case depends only on deterministic policies and purpose-owned driven ports. Runtime children, pinned successor JSON transport, and public-report projection are outbound adapters selected by the update registry composition root. |
 | `integrations/**` | external-system parsing and operations | Outbound adapters. |
 
 `index.ts` and `types.ts` are filenames, not roles. A pure `index.ts` barrel may
@@ -268,7 +269,7 @@ No single layer owns all truth.
 | Observer recovery inventory | Point-in-time retained-session and redacted recovery-handle evidence. It never classifies recovery eligibility or grants mutation authority. |
 | Observer recovery assessment | Point-in-time classification over one recovery inventory and one separately captured graph. It may select an opaque Observer handle for downstream policy but never launches it or grants mutation authority; public update reports omit that handle identity. |
 | Update aggregate inspection | One read-only CLI-owned aggregation of Observer identity/health/recovery, Host build/inventory, provider hook health, and terminal dispositions for one installed and selected artifact. One required port constructs the public aggregate and private sidecar together, then strict CLI-owned validation requires matching Observer build selectors and one unique exact opaque handle for every public selected-handle session before projecting public-only #665 evidence. Exact executable provenance is normally required. Atomic installation replacement has one narrower evidence state: exact parsed argv, pidfile, health, OS start, process token, build selector, and socket remain pinned while only the old live image differs from the currently installed path identity, so only the existing explicit same-or-higher-build cooperative restart is eligible. Missing or ambiguous process records, executable or argv mismatch, lower builds, automatic signals, and destructive recovery still refuse. Public evidence is redaction-safe; its private sidecar retains the Observer ownership tuple and selected opaque Station recovery-handle IDs only long enough to plan, digest, and revalidate. Neither form grants mutation authority. |
-| Update convergence plan | A deterministic provider-neutral decision over one aggregate inspection and explicit selected target. The evaluator (`incumbent-cli` or `successor-cli`) is separate from target ownership. Successor evaluators use the install owner's local exact-target proof, not ordinary latest discovery. Public phases and component decisions plus the digest are serialized in report v4; a completed terminal-handoff audit additionally carries only the canonical immutable PTY lifetime receipt. Process tokens, recovery-handle IDs, aggregate paths, raw provider data, and process groups are not serialized. Final verification after handoff requires the same exact terminal set; an absent terminal needs an explicit evidenced disposition and is never inferred to have exited naturally. A verified no-op plan is the sole source of `current`. A typed public policy sanitizes every nested SafeError after child parsing and defensively before JSON or text presentation. The separate `execution-failed` diagnostic envelope preserves only a strict sanitized Observer lifecycle cause and bounded startup evidence, including through a successor report. |
+| Update convergence plan | A deterministic provider-neutral decision over one aggregate inspection and explicit selected target. The evaluator (`incumbent-cli` or `successor-cli`) is separate from target ownership. Successor evaluators use the install owner's local exact-target proof, not ordinary latest discovery. Public phases and component decisions plus the digest are serialized in report v4; a completed terminal-handoff audit additionally carries only the canonical immutable PTY lifetime receipt. Process tokens, recovery-handle IDs, aggregate paths, raw provider data, and process groups are not serialized. Final verification after handoff requires the same exact terminal set; an absent terminal needs an explicit evidenced disposition and is never inferred to have exited naturally. A verified no-op plan is the sole source of `current`. A typed public-report adapter sanitizes every nested SafeError after child parsing and defensively before JSON or text presentation. Non-mutating text shows all seven result-phase statuses, and one terminal-safe encoder protects every untrusted rendered string. The separate `execution-failed` diagnostic envelope preserves only a strict sanitized Observer lifecycle cause and bounded startup evidence, including through a successor report. |
 | Update convergence digest | Versioned SHA-256 commitment to authorization-relevant target, aggregate, private ownership, and selected Station-handle facts under an explicit canonical encoding that rejects ill-formed Unicode before UTF-8 hashing. It reveals no private value, is not authority, and never substitutes for #641's fresh exact process-group authorization and journal. |
 
 Clients must treat a subscription gap as possible event loss. The Station client
